refactor(scripting_api): move DirectoryPayloadFormat to its own file
