+++ lib/src/core/scripting_api.dart
@@ -12,6 +12,7 @@ library scripting_api;
 
 export "scripting_api/consumed_thing.dart";
 export "scripting_api/data_schema_value.dart";
+export "scripting_api/discovery/directory_payload_format.dart";
 export "scripting_api/discovery/thing_discovery.dart";
 export "scripting_api/discovery/thing_filter.dart";
 export "scripting_api/exposed_thing.dart";
